optimise: optimise the code and modify the style
