+++ src/Components/FilterTasks/FilterTask.tsx
@@ -1,6 +1,6 @@
 import { UseFormRegister } from "react-hook-form";
 import "./FilterTask.css";
-import { FilterElement } from "../../types/types";
+import { FilterElement } from "../../types/TaskType/types";
 
 export default function FilterTask({
   register,
