--- src/context/TaskContext.tsx
@@ -1,11 +1,0 @@
-import { Task } from "../types/types";
-import { createContext } from "react";
-export type TasksArrayType = {
-  tasks: Task[];
-  setTasks: (c: Task[]) => void;
-};
-
-export const TaskContext = createContext<TasksArrayType>({
-  tasks: [],
-  setTasks: () => {},
-});
