+++ src/index.css
@@ -38,7 +38,8 @@ button {
   font-size: 1em;
   font-weight: 500;
   font-family: inherit;
-  background-color: #474747;
+  background-color: #242323;
+  color: white;
   cursor: pointer;
   transition: border-color 0.25s;
 }
